requirement.txt
skimage >= 0.11.0
scipy >= 0.15.1
numpy >= 1.11.1
h5py >= 2.7.0
PIL >= 1.1.7
absl-py >= 0.1.4
